refactor: restructure optimizer tab
