+++ src/lib/gpu/webgpuOptimizer.ts
@@ -8,7 +8,7 @@ import { OptimizerDisplayData } from 'lib/optimizer/bufferPacker'
 import { calculateBuild } from 'lib/optimizer/calculateBuild'
 import { renameFields } from 'lib/optimizer/optimizer'
 import { SortOption } from 'lib/optimizer/sortOptions'
-import { setSortColumn } from 'lib/tabs/tabOptimizer/optimizerForm/RecommendedPresetsButton'
+import { setSortColumn } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { Form } from 'types/form'
 import { OptimizerContext } from 'types/optimizer'
