+++ src/lib/overlays/modals/ScoringModal.tsx
@@ -4,7 +4,7 @@ import { Parts, Stats } from 'lib/constants/constants'
 import { Assets } from 'lib/rendering/assets'
 import { ScoringMetadata } from 'lib/scoring/characterScorer'
 import DB from 'lib/state/db'
-import CharacterSelect from 'lib/tabs/tabOptimizer/optimizerForm/CharacterSelect'
+import CharacterSelect from 'lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelect'
 import { ColorizedLinkWithIcon } from 'lib/ui/ColorizedLink'
 import { TsUtils } from 'lib/utils/TsUtils'
 import React, { useEffect } from 'react'
