+++ src/lib/scoring/characterScorer.ts
@@ -17,7 +17,7 @@ import {
   SimulationStats,
 } from 'lib/simulations/statSimulationController'
 import DB from 'lib/state/db'
-import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/StatSimulationDisplay'
+import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay'
 import { TsUtils } from 'lib/utils/TsUtils'
 import { Utils } from 'lib/utils/utils'
 import { Character } from 'types/character'
