+++ src/lib/simulations/statSimulationController.tsx
@@ -12,8 +12,8 @@ import { Assets } from 'lib/rendering/assets'
 import { SimulationResult } from 'lib/scoring/characterScorer'
 import DB from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
-import { setSortColumn } from 'lib/tabs/tabOptimizer/optimizerForm/RecommendedPresetsButton'
-import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/StatSimulationDisplay'
+import { setSortColumn } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
+import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { Utils } from 'lib/utils/utils'
 import { useTranslation } from 'react-i18next'
