+++ src/lib/state/db.ts
@@ -20,8 +20,8 @@ import { Themes } from 'lib/rendering/theme'
 import { ScoringMetadata, SimulationMetadata } from 'lib/scoring/characterScorer'
 import { oldCharacterScoringMetadata } from 'lib/scoring/oldCharacterScoringMetadata'
 import { SaveState } from 'lib/state/saveState'
-import { OptimizerMenuIds } from 'lib/tabs/tabOptimizer/FormRow'
-import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/StatSimulationDisplay'
+import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay'
+import { OptimizerMenuIds } from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormRow'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { TsUtils } from 'lib/utils/TsUtils'
 import { Utils } from 'lib/utils/utils'
