+++ src/lib/tabs/tabOptimizer/OptimizerForm.tsx
@@ -7,22 +7,22 @@ import DB from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
 import { CharacterConditionalsDisplay } from 'lib/tabs/tabOptimizer/conditionals/CharacterConditionalsDisplay'
 import { LightConeConditionalDisplay } from 'lib/tabs/tabOptimizer/conditionals/LightConeConditionalDisplay'
-import FilterContainer from 'lib/tabs/tabOptimizer/FilterContainer'
-import FormCard from 'lib/tabs/tabOptimizer/FormCard'
-import { FormRow, OptimizerMenuIds, TeammateFormRow } from 'lib/tabs/tabOptimizer/FormRow'
-import { AdvancedOptionsPanel } from 'lib/tabs/tabOptimizer/optimizerForm/AdvancedOptionsPanel'
-import CharacterSelectorDisplay from 'lib/tabs/tabOptimizer/optimizerForm/CharacterSelectorDisplay'
-import { CombatBuffsDrawer } from 'lib/tabs/tabOptimizer/optimizerForm/CombatBuffsDrawer'
-import { ComboFilters } from 'lib/tabs/tabOptimizer/optimizerForm/ComboFilter'
-import { EnemyConfigurationsDrawer } from 'lib/tabs/tabOptimizer/optimizerForm/EnemyConfigurationsDrawer'
-import { FormSetConditionals } from 'lib/tabs/tabOptimizer/optimizerForm/FormSetConditionals'
-import OptimizerOptionsDisplay from 'lib/tabs/tabOptimizer/optimizerForm/OptimizerOptionsDisplay'
-import { OptimizerTabCharacterPanel } from 'lib/tabs/tabOptimizer/optimizerForm/OptimizerTabCharacterPanel'
-import RelicMainSetFilters from 'lib/tabs/tabOptimizer/optimizerForm/RelicMainSetFilters'
-import { MinMaxRatingFilters, MinMaxStatFilters } from 'lib/tabs/tabOptimizer/optimizerForm/ResultFilters'
-import { StatSimulationDisplay } from 'lib/tabs/tabOptimizer/optimizerForm/StatSimulationDisplay'
-import { SubstatWeightFilters } from 'lib/tabs/tabOptimizer/optimizerForm/SubstatWeightFilters'
-import TeammateCard from 'lib/tabs/tabOptimizer/optimizerForm/TeammateCard'
+import { AdvancedOptionsPanel } from 'lib/tabs/tabOptimizer/optimizerForm/components/AdvancedOptionsPanel'
+import CharacterSelectorDisplay from 'lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelectorDisplay'
+import { CombatBuffsDrawer } from 'lib/tabs/tabOptimizer/optimizerForm/components/CombatBuffsDrawer'
+import { ComboFilters } from 'lib/tabs/tabOptimizer/optimizerForm/components/ComboFilter'
+import { EnemyConfigurationsDrawer } from 'lib/tabs/tabOptimizer/optimizerForm/components/EnemyConfigurationsDrawer'
+import { FormSetConditionals } from 'lib/tabs/tabOptimizer/optimizerForm/components/FormSetConditionals'
+import OptimizerOptionsDisplay from 'lib/tabs/tabOptimizer/optimizerForm/components/OptimizerOptionsDisplay'
+import { OptimizerTabCharacterPanel } from 'lib/tabs/tabOptimizer/optimizerForm/components/OptimizerTabCharacterPanel'
+import RelicMainSetFilters from 'lib/tabs/tabOptimizer/optimizerForm/components/RelicMainSetFilters'
+import { MinMaxRatingFilters, MinMaxStatFilters } from 'lib/tabs/tabOptimizer/optimizerForm/components/ResultFilters'
+import { StatSimulationDisplay } from 'lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay'
+import { SubstatWeightFilters } from 'lib/tabs/tabOptimizer/optimizerForm/components/SubstatWeightFilters'
+import TeammateCard from 'lib/tabs/tabOptimizer/optimizerForm/components/TeammateCard'
+import FilterContainer from 'lib/tabs/tabOptimizer/optimizerForm/layout/FilterContainer'
+import FormCard from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormCard'
+import { FormRow, OptimizerMenuIds, TeammateFormRow } from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormRow'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { Utils } from 'lib/utils/utils'
 import React, { useEffect } from 'react'
