+++ src/lib/tabs/tabOptimizer/OptimizerTab.tsx
@@ -3,7 +3,7 @@ import 'ag-grid-community/styles/ag-grid.css'
 import OptimizerBuildPreview from 'lib/tabs/tabOptimizer/OptimizerBuildPreview'
 
 import OptimizerForm from 'lib/tabs/tabOptimizer/OptimizerForm'
-import { OptimizerGrid } from 'lib/tabs/tabOptimizer/OptimizerGrid'
+import { OptimizerGrid } from 'lib/tabs/tabOptimizer/optimizerForm/grid/OptimizerGrid'
 import { ZeroPermutationsSuggestionsModal, ZeroResultSuggestionModal } from 'lib/tabs/tabOptimizer/OptimizerSuggestionsModal'
 import Sidebar from 'lib/tabs/tabOptimizer/Sidebar'
 import React from 'react'
