+++ src/lib/tabs/tabOptimizer/Sidebar.tsx
@@ -9,7 +9,7 @@ import { Message } from 'lib/interactions/message'
 import { Optimizer } from 'lib/optimizer/optimizer'
 import { SettingOptions } from 'lib/overlays/drawers/SettingsDrawer'
 import DB from 'lib/state/db'
-import { defaultPadding } from 'lib/tabs/tabOptimizer/optimizerTabConstants'
+import { defaultPadding } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridConstants'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { HeaderText } from 'lib/ui/HeaderText'
 import { TooltipImage } from 'lib/ui/TooltipImage'
