+++ src/lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelect.tsx
@@ -1,6 +1,6 @@
 import { Button, Card, Flex, Input, InputRef, Modal, Select } from 'antd'
 import { Assets } from 'lib/rendering/assets'
-import { CardGridItemContent, generateElementTags, generatePathTags, SegmentedFilterRow } from 'lib/tabs/tabOptimizer/optimizerForm/CardSelectModalComponents'
+import { CardGridItemContent, generateElementTags, generatePathTags, SegmentedFilterRow } from 'lib/tabs/tabOptimizer/optimizerForm/components/CardSelectModalComponents'
 import { Utils } from 'lib/utils/utils'
 import * as React from 'react'
 import { useEffect, useMemo, useRef, useState } from 'react'
