+++ src/lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelectorDisplay.tsx
@@ -1,10 +1,10 @@
 import { Flex, Form, Select } from 'antd'
 import { Hint } from 'lib/interactions/hint'
 import { SortOption } from 'lib/optimizer/sortOptions'
-import CharacterSelect from 'lib/tabs/tabOptimizer/optimizerForm/CharacterSelect'
-import LightConeSelect from 'lib/tabs/tabOptimizer/optimizerForm/LightConeSelect'
-import RecommendedPresetsButton from 'lib/tabs/tabOptimizer/optimizerForm/RecommendedPresetsButton'
-import { optimizerTabDefaultGap, panelWidth } from 'lib/tabs/tabOptimizer/optimizerTabConstants'
+import CharacterSelect from 'lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelect'
+import LightConeSelect from 'lib/tabs/tabOptimizer/optimizerForm/components/LightConeSelect'
+import RecommendedPresetsButton from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
+import { optimizerTabDefaultGap, panelWidth } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridConstants'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { HeaderText } from 'lib/ui/HeaderText'
 import { TooltipImage } from 'lib/ui/TooltipImage'
