+++ src/lib/tabs/tabOptimizer/optimizerForm/components/CombatBuffsDrawer.tsx
@@ -1,8 +1,8 @@
 import { Drawer, Flex, Form, Typography } from 'antd'
 import { CombatBuffs } from 'lib/constants/constants'
 import { defaultGap } from 'lib/constants/constantsUi'
-import InputNumberStyled from 'lib/tabs/tabOptimizer/optimizerForm/InputNumberStyled'
-import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerTabConstants'
+import InputNumberStyled from 'lib/tabs/tabOptimizer/optimizerForm/components/InputNumberStyled'
+import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridConstants'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
