+++ src/lib/tabs/tabOptimizer/optimizerForm/components/LightConeSelect.tsx
@@ -1,7 +1,7 @@
 import { Card, Flex, Input, InputRef, Modal, Select } from 'antd'
 import { Assets } from 'lib/rendering/assets'
 import DB from 'lib/state/db'
-import { CardGridItemContent, generatePathTags, generateRarityTags, SegmentedFilterRow } from 'lib/tabs/tabOptimizer/optimizerForm/CardSelectModalComponents'
+import { CardGridItemContent, generatePathTags, generateRarityTags, SegmentedFilterRow } from 'lib/tabs/tabOptimizer/optimizerForm/components/CardSelectModalComponents'
 import { Utils } from 'lib/utils/utils'
 import * as React from 'react'
 import { useEffect, useMemo, useRef, useState } from 'react'
