+++ src/lib/tabs/tabOptimizer/optimizerForm/components/OptimizerTabCharacterPanel.tsx
