+++ src/lib/tabs/tabOptimizer/optimizerForm/components/ResultFilters.tsx
@@ -1,7 +1,7 @@
 import { Flex } from 'antd'
 import { Hint } from 'lib/interactions/hint'
-import FilterRow from 'lib/tabs/tabOptimizer/optimizerForm/FilterRow'
-import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerTabConstants'
+import FilterRow from 'lib/tabs/tabOptimizer/optimizerForm/layout/FilterRow'
+import { optimizerTabDefaultGap } from 'lib/tabs/tabOptimizer/optimizerForm/grid/optimizerGridConstants'
 import { HeaderText } from 'lib/ui/HeaderText'
 import { TooltipImage } from 'lib/ui/TooltipImage'
 import { useTranslation } from 'react-i18next'
