+++ src/lib/tabs/tabOptimizer/optimizerForm/components/SimulatedBuildsGrid.tsx
@@ -2,7 +2,7 @@ import { CloseOutlined } from '@ant-design/icons'
 import { IRowNode } from 'ag-grid-community'
 import { Empty, Flex, Table, TableColumnsType } from 'antd'
 import { deleteStatSimulationBuild, renderDefaultSimulationName } from 'lib/simulations/statSimulationController'
-import { STAT_SIMULATION_GRID_WIDTH } from 'lib/tabs/tabOptimizer/optimizerForm/StatSimulationDisplay'
+import { STAT_SIMULATION_GRID_WIDTH } from 'lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay'
 import { Utils } from 'lib/utils/utils'
 import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
