+++ src/lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay.tsx
@@ -3,11 +3,11 @@ import { Button, Flex, Form, Input, InputNumber, Popconfirm, Radio, Select, Typo
 import { Parts, Stats, SubStats } from 'lib/constants/constants'
 import { Assets } from 'lib/rendering/assets'
 import { deleteAllStatSimulationBuilds, importOptimizerBuild, saveStatSimulationBuildFromForm, startOptimizerStatSimulation } from 'lib/simulations/statSimulationController'
-import FormCard from 'lib/tabs/tabOptimizer/FormCard'
-import { OrnamentSetTagRenderer } from 'lib/tabs/tabOptimizer/optimizerForm/OrnamentSetTagRenderer'
-import GenerateOrnamentsOptions from 'lib/tabs/tabOptimizer/optimizerForm/OrnamentsOptions'
-import { GenerateBasicSetsOptions } from 'lib/tabs/tabOptimizer/optimizerForm/SetsOptions'
-import { SimulatedBuildsGrid } from 'lib/tabs/tabOptimizer/optimizerForm/SimulatedBuildsGrid'
+import { OrnamentSetTagRenderer } from 'lib/tabs/tabOptimizer/optimizerForm/components/OrnamentSetTagRenderer'
+import GenerateOrnamentsOptions from 'lib/tabs/tabOptimizer/optimizerForm/components/OrnamentsOptions'
+import { GenerateBasicSetsOptions } from 'lib/tabs/tabOptimizer/optimizerForm/components/SetsOptions'
+import { SimulatedBuildsGrid } from 'lib/tabs/tabOptimizer/optimizerForm/components/SimulatedBuildsGrid'
+import FormCard from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormCard'
 import { VerticalDivider } from 'lib/ui/Dividers'
 import { HeaderText } from 'lib/ui/HeaderText'
 import { Utils } from 'lib/utils/utils'
