+++ src/lib/tabs/tabOptimizer/optimizerForm/components/TeammateCard.tsx
@@ -8,9 +8,9 @@ import { Assets } from 'lib/rendering/assets'
 import DB from 'lib/state/db'
 import { CharacterConditionalsDisplay } from 'lib/tabs/tabOptimizer/conditionals/CharacterConditionalsDisplay'
 import { LightConeConditionalDisplay } from 'lib/tabs/tabOptimizer/conditionals/LightConeConditionalDisplay'
-import FormCard from 'lib/tabs/tabOptimizer/FormCard'
-import CharacterSelect from 'lib/tabs/tabOptimizer/optimizerForm/CharacterSelect'
-import LightConeSelect from 'lib/tabs/tabOptimizer/optimizerForm/LightConeSelect'
+import CharacterSelect from 'lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelect'
+import LightConeSelect from 'lib/tabs/tabOptimizer/optimizerForm/components/LightConeSelect'
+import FormCard from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormCard'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
