+++ src/lib/tabs/tabOptimizer/optimizerTabController.ts
@@ -14,7 +14,7 @@ import { GridAggregations } from 'lib/rendering/gradient'
 import { CharacterStats } from 'lib/scoring/characterStats'
 import DB from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
-import { applyMetadataPresetToForm } from 'lib/tabs/tabOptimizer/optimizerForm/RecommendedPresetsButton'
+import { applyMetadataPresetToForm } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
 import { TsUtils } from 'lib/utils/TsUtils'
 import { Utils } from 'lib/utils/utils'
 import { Form, Teammate } from 'types/form'
