+++ src/lib/tabs/tabShowcase/RelicScorerTab.jsx
@@ -17,7 +17,7 @@ import CharacterModal from 'lib/overlays/modals/CharacterModal'
 import { Assets } from 'lib/rendering/assets'
 import DB, { AppPages, PageToRoute } from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
-import { applySpdPreset } from 'lib/tabs/tabOptimizer/optimizerForm/RecommendedPresetsButton'
+import { applySpdPreset } from 'lib/tabs/tabOptimizer/optimizerForm/components/RecommendedPresetsButton'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
 import { Utils } from 'lib/utils/utils'
 import PropTypes from 'prop-types'
