+++ src/types/store.d.ts
@@ -2,7 +2,7 @@ import { ComputeEngine } from 'lib/constants/constants'
 import { ComboState } from 'lib/optimizer/rotation/comboDrawerController'
 import { ColorThemeOverrides } from 'lib/rendering/theme'
 import { ScoringMetadata } from 'lib/scoring/characterScorer'
-import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/StatSimulationDisplay'
+import { StatSimTypes } from 'lib/tabs/tabOptimizer/optimizerForm/components/StatSimulationDisplay'
 import { Build, Character, CharacterId } from 'types/character'
 import { Form } from 'types/form'
 import { Relic } from 'types/relic'
